// nodes/ZohoCliq/v1/actions/node.type.ts
/**
 * Type unions for Zoho Cliq node resources and operations
 */

export type ZohoCliqType =
	| 'message'
	| 'messageComponentBuilder'
	| 'oauthHelper'
	| 'channel'
	| 'bot'
	| 'thread'
	| 'chat'
	| 'file'
	| 'reaction'
	| 'user'
	| 'team'
	| 'department'
	| 'role'
	| 'designation'
	| 'userField'
	| 'userStatus'
	| 'event'
	| 'reminder'
	| 'remoteWork'
	| 'callsMeeting'
	| 'widgetMapTicker'
	| 'customEmail'
	| 'database'
	| 'customDomain'
	| 'bulkAction';